requirements.txt
Django==1.8.1
Pillow==2.9.0
PyJWT==1.3.0
defusedxml==0.4.1
django-appconf==1.0.1
django-bootstrap3==6.1.0
django-compressor==1.5
django-debug-toolbar==1.3.2
django-extensions==1.5.6
django-registration==1.0
ipython==3.2.0
lxml==3.4.4
oauthlib==0.7.2
python-social-auth==0.2.11
python3-openid==3.0.5
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
sorl-thumbnail==12.3
sqlparse==0.1.15
